+++ packages/terra-functional-testing/CHANGELOG.md
@@ -5,7 +5,11 @@
 * Added
   * Added authentication for accessing screenshots from the remote site.
   * Added `useRemoteReferenceScreenshots` cli option for downloading reference screenshots from a remote site for screenshot comparisons.
+  * Delete the `reference` screenshot directory when `useRemoteReferenceScreenshots` is true.
   * Added a check of the `BUILD_BRANCH` environment variable to determines if tests should pass regardless of image mismatch
+
+* Fixed
+  * Fixed deleting the `diff`, `error`, and optionally the `latest` screenshot directories at the beginning of each test run.
 
 ## 2.7.0 - (February 11, 2022)